requirements.txt
crewai>=0.28.8
python-pptx>=0.6.21
python-dotenv>=1.0.0
perplexity-api>=0.1.3
requests>=2.31.0
streamlit>=1.28.0